Build fixes for !ENABLE(SERVICE_WORKER) after 260990@main https://bugs.webkit.org/show_bug.cgi?id=253187
Reviewed by Alex Christensen.

Add ENABLE(SERVICE_WORKER) guards to code in OriginStorageManager
that is using StorageType::BackgroundFetchStorage.

We saw errors where the generated serializers appeared to be effectively
missing the trailing semi-colon on the allValidBitsValue calculation
in isValidOptionSet due to the last item being in a preprocessor if.
Add a check for the case where the last item is in such an if and add an
else that provides the semi-colon.

* Source/WebKit/NetworkProcess/storage/OriginStorageManager.cpp:
(WebKit::OriginStorageManager::StorageBucket::toStorageIdentifier const):
(WebKit::OriginStorageManager::StorageBucket::resolvedPath):
* Source/WebKit/Scripts/generate-serializers.py:
(generate_impl):
* Source/WebKit/Scripts/webkit/tests/GeneratedSerializers.cpp:
* Source/WebKit/Scripts/webkit/tests/GeneratedSerializers.h:
* Source/WebKit/Scripts/webkit/tests/SerializedTypeInfo.cpp:
* Source/WebKit/Scripts/webkit/tests/TestSerializedType.serialization.in:

Canonical link: https://commits.webkit.org/261094@main