Revert 256469@main as it violates the specification https://bugs.webkit.org/show_bug.cgi?id=253435 rdar://106273441
Reviewed by Alex Christensen.

Revert 256469@main. The specification [1] says that we should not expose this
API on non-mobile devices. This broke the Final Cut Pro download page as it
relies on this API being missing to detect we're on desktop.

Note that neither Firefox not Chrome expose the window orientation API on
Desktop.

[1] https://compat.spec.whatwg.org/#windoworientation-interface

* LayoutTests/imported/w3c/web-platform-tests/compat/idlharness.window-expected.txt:
* Source/WTF/wtf/PlatformEnable.h:
* Source/WTF/wtf/PlatformEnableCocoa.h:
* Source/WebCore/html/HTMLBodyElement+Compat.idl:
* Source/WebCore/page/DOMWindow+Compat.idl:
* Source/WebCore/page/DOMWindow.cpp:
(WebCore::DOMWindow::orientation const):
* Source/WebCore/page/DOMWindow.h:

Canonical link: https://commits.webkit.org/261282@main