Add API tests for "consume gesture if not verifably from UI process" policy https://bugs.webkit.org/show_bug.cgi?id=253479 rdar://106340542
Reviewed by NOBODY (OOPS!).

In https://commits.webkit.org/261190@main, a policy to consume user gestures if
not verifably from the UI process was introduced for window.open behind a flag. This patch adds
API tests to ensure this policy is followed when reporting user gestures to the
client.

* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/WebKit/open-window-button.html: Added.
* Tools/TestWebKitAPI/Tests/WebKitCocoa/OpenAndCloseWindow.mm:
(resetToConsistentState):
(-[OpenWindowUIDelegate webView:createWebViewWithConfiguration:forNavigationAction:windowFeatures:]):
(TEST):