imported/w3c/web-platform-tests/background-fetch/abort.https.window.html is flaky https://bugs.webkit.org/show_bug.cgi?id=253314 rdar://problem/106195333
Reviewed by NOBODY (OOPS!).

The test was flakily crashing in debug bots when trying to resolve the promise of BackgroundFetchRegistration::matchAll.
The issue is that the toJS call fails due to an exception.
It returns an empty value that is used to resolve the promise.

To prevent this, we add a check in DeferredPromise::callFunction to exit early in that case.

* Source/WebCore/bindings/js/JSDOMPromiseDeferred.cpp:
(WebCore::DeferredPromise::callFunction):