Mark layout needed on floats that have trimmed margins in block layout https://bugs.webkit.org/show_bug.cgi?id=252183 rdar://105410985
Reviewed by NOBODY (OOPS!).

During block layout we attempt to find a location for floats and trim
its margin if needed to when we find a suitable position for it. This
is done by removing the margin in the renderer's m_marginBox and also
adjusting the associated FloatingObject's logical width by removing
the trimmed margin from it since the logical width is the margin box
logical width in that case. However, if a relayout occurs and the
block container's width remains the same as the one that was computed
from its previous layout, then we may position the float's incorrectly
during this new iteration of layout. For example:

<style>
.container {
    width: 100px;
    margin-trim: inline;
}
</style>
<div class="container">
  <div style="width: 100px; height: 10px; background-color: green"></div>
  <div style="width: 25px; height: 50px; float: left; background-color: blue; margin-inline-start: 20px;"></div>
  <div style="width: 25px; height: 50px; float: right; height: 80px; background-color: gray; margin-inline-end: 20px;"></div>
  <div style="width: 25px; height: 50px; float: left; background-color: purple; margin-inline-start: 80px;"></div>
</div>

When layout is finished, this may look something like the following:

----------------------------------------
|   green div at the top               |
|                                      |
----------------------------------------
-------------            ---------------
| blue float|            | gray float  |
-------------            |             |
--------------           |             |
|purple float|           |             |
--------------           |             |
                         ---------------

In this scenario all of the floats have had their margins trimmed when
they got to their final positions next to the containing box's inner
edge. However, if a relayout occurs (for example by setting the root
HTML element to a width of 99%) then the purple float will get positioned
incorreclty in between the blue and gray one. This is because
RenderBlockFlow::layoutBlockChildren will have its relayoutChildren
argument set to false (since the width of the containing block did not
change from the previous layout). So as it iterats over each of its
children and calls updateBlockChildDirtyBitsBeforeLayout, the children
will not get marked for layout. By the time we get to positioning the
floats in RenderBlockFlow::computeLogicalLocationForFloat we will be
using the float's trimmed dimensions during positioning and will place
it incorrectly. By adding an extra condition inside of  RenderBlock::updateBlockChildDirtyBitsBeforeLayout
and checking to see if the child is a float with a trimmed margin, we
can position the floats using their original dimensions.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/block-container-overconstraining-float-margins-tirmmed-at-final-position-block-layout-expected.html: Renamed from LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/block-container-overflowing-float-margins-tirmmed-at-final-position-block-layout-expected.html.
* LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/block-container-overconstraining-float-margins-tirmmed-at-final-position-block-layout.html: Renamed from LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/block-container-overflowing-float-margins-tirmmed-at-final-position-block-layout.html.
* Source/WebCore/rendering/RenderBlock.cpp:
(WebCore::RenderBlock::childIsFloatWithTrimmedMargin const):
(WebCore::RenderBlock::updateBlockChildDirtyBitsBeforeLayout):
* Source/WebCore/rendering/RenderBlock.h: