[GTK] Test fast/text/font-promises-gc.html fails after r281465 https://bugs.webkit.org/show_bug.cgi?id=229732
Reviewed by Alan Baradlay.

The test was failing for WebKitGTK/WPE because a console error message
was different. Font was 'unloaded' for WebKitGTK/WPE but 'loaded' for
general baseline. This message is informative and it doesn't affect the
result of the test, which was passing too for WebKitGTK/WPE.

* LayoutTests/fast/text/font-promises-gc-expected.txt:
* LayoutTests/fast/text/font-promises-gc.html:

Canonical link: https://commits.webkit.org/261364@main