[git-webkit] Use fsmonitor and untrackedCache by default https://bugs.webkit.org/show_bug.cgi?id=253586 rdar://106439081
Reviewed by NOBODY (OOPS!).

The git index for the webkit repo is huge, and most git commands (such as status and diff)
will refresh the index before doing anything else, which requires hundreds of thousands
of lstat calls. Instead, we can use the `core.fsmonitor` config to use a daemon to monitor
the file system as describe in details in the github blog[1]. Local testing shows a reduction
in the time `git status` takes from over 10s to less than 100ms.

[1]: https://github.blog/2022-06-29-improve-git-monorepo-performance-with-a-file-system-monitor/

* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/setup.py:
(Setup.git):